# [Refactor] Explode PgFramestamp fields