… README.md
… lib/ecto/postgres/pg_framestamp_range_migrations.ex
… lib/ecto/postgres/pg_rational_migrations.ex
… lib/ecto/postgres/pg_utils.ex
… lib/framestamp.ex
… lib/framestamp_mixed_rate_arithmatic_error.ex
… lib/framestamp_range.ex
… lib/framestamp_range_mixed_out_type_arithmatic_error.ex
… lib/utils/mixed_rate_ops.ex
… lib/utils/parse.ex
… test/ecto/postgres/pg_framestamp_test.exs
… test/ecto/postgres/pg_rational_test.exs
… test/framestamp/framestamp_ops_test.exs
… test/framestamp_range_test.exs
… zdocs/ecto_quickstart.cheatmd
… zdocs/framerate_vs_timebase.md
… zdocs/the_rational_rationale.md