Flattens the data in `PgFramestamp` to avoid nested composites. This change offers SIGNIFICANT performance improvements

## Overall results

|    TEST                     | Ops  | Before   |  After   | Speedup | Sep. Fields | Speedup |
| --------------------------- | ---- | -------: | -------: | ------: | ----------: | ------: |
| Inserts                     | 1M   |  3670 ms |  2722 ms | 1.3x    | 2361 ms     | 0.9x    |
| Inserts w/ constraints      | 1M   | 21602 ms | 15768 ms | 1.4x    |             |         |
| Inserts w/ const. & indexes | 100T | 74786 ms | 36958 ms | 2.0x    |             |         |
| Aritmatic Operations        | 1M   |  2092 ms |   741 ms | 2.8x    | 727 ms      | 1.0x    |
| Comparisons                 | 1M   |  1034 ms |   518 ms | 2.0x    |             |         |
| Overlap Joins w/indexes     | 100T |   325 ms |   197 ms | 1.6x    |             |         |

One additional note, a test was run for doing 1M `bigint` sums on a record with 8 `bigint` fields, and that operation yielded `1378 ms`, making the two-field framestamp 1M sum operation `1.9x` faster, indicating that we get BIG gains by keeping the number of fields down, and are within an order of magnitude using these custom operators.

Another test was done using 3 double precision float fields, which was able to compare 1M float fields in `139 ms`, meaning that our rational framestamp comparison is only `3.8x` slower. Overall, that's VERY good given the number of operations that have to happen in order to add rational numbers together as opposed to floats.

## NEW TABLES USED

```sql
CREATE TABLE triple_rationals (
    id uuid,
    a_seconds rational,
    b_seconds rational,
    rate rational,
    rate_tags framerate_tags[]
);

CREATE TABLE framestamps_03 (
    id uuid,
    a framestamp,
    b framestamp
);
```

## INSERTS -- NO CONSTRAINTS OR INDEXES

#### BEFORE | Framestamp

```sql
INSERT INTO framestamps_03 (id, a, b)
SELECT 
    gen_random_uuid(),
    (
        (generate_series(1, 1000000), 24), 
        ((24, 1), '{}')
    )::framestamp,
    (
        (generate_series(1000001, 2000000), 24), 
        ((24, 1), '{}')
    )::framestamp;
```

```text
Time: 3670.154 ms (00:03.670)
```

#### AFTER | Framestamp

```sql
INSERT INTO framestamps_03 (id, a, b)
SELECT 
    gen_random_uuid(),
    (
        generate_series(1, 1000000),
        24,
        24,
        1,
        '{}'
    )::framestamp,
    (
        generate_series(1000001, 2000000),
        24,
        24,
        1,
        '{}'
    )::framestamp;
```

```text
Time: 2722.012 ms (00:02.722)
```

#### AFTER | Exploded Rational Fields

```sql
INSERT INTO triple_rat_table (id, a_seconds, b_seconds, rate, tags)
SELECT
    gen_random_uuid(),
    (
        generate_series(1, 1000000),
	24
    )::rational,
    (
        generate_series(1, 1000000),
	24
    )::rational,
    (24, 1)::rational,
    '{}';
```

```text
Time: 2731.476 ms (00:02.731)
```

## INSERTS -- With Constraints

#### Before

```sql
INSERT INTO framestamps_01 (id, a, b)
SELECT 
    gen_random_uuid(),
    (
        (generate_series(1, 1000000), 24), 
        ((24, 1), '{}')
    )::framestamp,
    (
        (generate_series(1000001, 2000000), 24), 
        ((24, 1), '{}')
    )::framestamp;
```

```text
Time: 21602.279 ms (00:21.602)
```

#### After

```sql
INSERT INTO framestamps_01 (id, a, b)
SELECT 
    gen_random_uuid(),
    (
        generate_series(1, 1000000),
        24,
        24,
        1,
        '{}'
    )::framestamp,
    (
        generate_series(1000001, 2000000),
        24,
        24,
        1,
        '{}'
    )::framestamp;
```

```text
Time: 15767.785 ms (00:15.768)
```

## INSERTS -- With Indexes & Constraints

#### Before

```sql
INSERT INTO framestamps_02 (id, a, b)
SELECT 
    gen_random_uuid(),
    (
        (generate_series(1, 100000), 24), 
        ((24, 1), '{}')
    )::framestamp,
    (
        (generate_series(100001, 200000), 24), 
        ((24, 1), '{}')
    )::framestamp;
```

```text
Time: 74785.819 ms (01:14.786)
```

#### After

```sql
INSERT INTO framestamps_02 (id, a, b)
SELECT 
    gen_random_uuid(),
    (
        generate_series(1, 100000),
        24,
        24,
        1,
        '{}'
    )::framestamp,
    (
        generate_series(100001, 200000),
        24,
        24,
        1,
        '{}'
    )::framestamp;
```

```
Time: 36958.203 ms (00:36.958)
```

## Aritmetic

#### Before

```sql
INSERT INTO framestamps_03 (id, a, b)
SELECT 
    gen_random_uuid(),
    (
        (generate_series(1, 1000000), 24), 
        ((24, 1), '{}')
    )::framestamp,
    (
        (generate_series(1000001, 2000000), 24), 
        ((24, 1), '{}')
    )::framestamp;

EXPLAIN ANALYZE SELECT a + b FROM framestamps_03;
```

```text
 Gather  (cost=1000.00..666000.00 rows=3000000 width=32) (actual time=97.067..2024.838 rows=1000000 loops=1)
   Workers Planned: 2
   Workers Launched: 2
   ->  Parallel Seq Scan on framestamps_03  (cost=0.00..365000.00 rows=1250000 width=32) (actual time=79.636..1912.558 rows=333333 loops=3)
 Planning Time: 0.156 ms
 JIT:
   Functions: 6
   Options: Inlining true, Optimization true, Expressions true, Deforming true
   Timing: Generation 0.805 ms, Inlining 152.755 ms, Optimization 40.917 ms, Emission 38.845 ms, Total 233.321 ms
 Execution Time: 2092.627 ms
```

#### After

```sql
INSERT INTO framestamps_03 (id, a, b)
SELECT 
    gen_random_uuid(),
    (
        generate_series(1, 1000000),
        24,
        24,
        1,
        '{}'
    )::framestamp,
    (
        generate_series(1000001, 2000000),
        24,
        24,
        1,
        '{}'
    )::framestamp;

EXPLAIN ANALYZE SELECT a + b FROM framestamps_03;
```

```text
 Gather  (cost=1000.00..370457.44 rows=1666725 width=32) (actual time=8.360..913.166 rows=1000000 loops=1)
   Workers Planned: 2
   Workers Launched: 2
   ->  Parallel Seq Scan on framestamps_03  (cost=0.00..202784.94 rows=694469 width=32) (actual time=5.231..809.671 rows=333333 loops=3)
 Planning Time: 0.255 ms
 JIT:
   Functions: 6
   Options: Inlining false, Optimization false, Expressions true, Deforming true
   Timing: Generation 0.790 ms, Inlining 0.000 ms, Optimization 1.347 ms, Emission 10.191 ms, Total 12.328 ms
 Execution Time: 988.935 ms
```

#### AFTER | Triple Rationals

```sql
INSERT INTO triple_rat_table (id, a_seconds, b_seconds, rate, tags)
SELECT
    gen_random_uuid(),
    (
	    generate_series(1, 1000000),
	    24
    )::rational,
    (
	    generate_series(1, 1000000),
	    24
    )::rational,
    (24, 1)::rational,
    '{}';

EXPLAIN ANALYZE SELECT a_seconds + b_seconds FROM triple_rationals;
```

```text
 Gather  (cost=1000.00..230614.40 rows=1000019 width=32) (actual time=7.601..652.781 rows=1000000 loops=1)
   Workers Planned: 2
   Workers Launched: 2
   ->  Parallel Seq Scan on triple_rationals  (cost=0.00..129612.50 rows=416675 width=32) (actual time=5.274..543.662 rows=333333 loops=3)
 Planning Time: 0.231 ms
 JIT:
   Functions: 6
   Options: Inlining false, Optimization false, Expressions true, Deforming true
   Timing: Generation 0.791 ms, Inlining 0.000 ms, Optimization 1.639 ms, Emission 9.288 ms, Total 11.719 ms
 Execution Time: 726.524 ms
(10 rows)
```

#### INT Table

```sql
INSERT INTO bigint_table (id, a, b, c, d, e, f, g, h, tags_a, tags_b)
SELECT
    gen_random_uuid(),
    generate_series(1, 1000000),
    generate_series(1000001, 2000000),
    generate_series(1, 1000000),
    generate_series(1000001, 2000000),
    generate_series(1, 1000000),
    generate_series(1000001, 2000000),
    generate_series(1, 1000000),
    generate_series(1000001, 2000000),
    '{}',
    '{}';

SELECT a + d FROM bigint_table;
```

```text
Time: 1377.712 ms (00:01.378)
```

## Comparison

#### Before

```sql
INSERT INTO framestamps_03 (id, a, b)
SELECT 
    gen_random_uuid(),
    (
        (generate_series(1, 1000000), 24), 
        ((24, 1), '{}')
    )::framestamp,
    (
        (generate_series(1000001, 2000000), 24), 
        ((24, 1), '{}')
    )::framestamp;

EXPLAIN ANALYZE SELECT a > b FROM framestamps_03;
```

```text
 Gather  (cost=1000.00..666000.00 rows=3000000 width=1) (actual time=72.207..964.967 rows=1000000 loops=1)
   Workers Planned: 2
   Workers Launched: 2
   ->  Parallel Seq Scan on framestamps_03  (cost=0.00..365000.00 rows=1250000 width=1) (actual time=65.650..851.548 rows=333333 loops=3)
 Planning Time: 0.271 ms
 JIT:
   Functions: 6
   Options: Inlining true, Optimization true, Expressions true, Deforming true
   Timing: Generation 0.742 ms, Inlining 127.673 ms, Optimization 39.443 ms, Emission 26.747 ms, Total 194.605 ms
 Execution Time: 1033.950 ms
```

#### After

```sql
INSERT INTO framestamps_03 (id, a, b)
SELECT 
    gen_random_uuid(),
    (
        generate_series(1, 1000000),
        24,
        24,
        1,
        '{}'
    )::framestamp,
    (
        generate_series(1000001, 2000000),
        24,
        24,
        1,
        '{}'
    )::framestamp;

EXPLAIN ANALYZE SELECT a > b FROM framestamps_03;
```

```text
 Gather  (cost=1000.00..370457.44 rows=1666725 width=1) (actual time=6.233..489.990 rows=1000000 loops=1)
   Workers Planned: 2
   Workers Launched: 2
   ->  Parallel Seq Scan on framestamps_03  (cost=0.00..202784.94 rows=694469 width=1) (actual time=5.638..368.069 rows=333333 loops=3)
 Planning Time: 0.093 ms
 JIT:
   Functions: 6
   Options: Inlining false, Optimization false, Expressions true, Deforming true
   Timing: Generation 2.533 ms, Inlining 0.000 ms, Optimization 2.141 ms, Emission 10.729 ms, Total 15.403 ms
 Execution Time: 518.801 ms
```

#### FLOAT TABLE

```sql
CREATE TABLE dual_floats (
        id uuid,
    a_seconds double precision,
    b_seconds double precision,
    rate double precision,
    tags framerate_tags[]

);

INSERT INTO dual_floats (id, a_seconds, b_seconds, rate, tags)
SELECT
    gen_random_uuid(),
    generate_series(1.0, 1000000.0),
    generate_series(1.0, 1000000.0),
    24.0,
    '{}';

EXPLAIN ANALYZE SELECT a_seconds + b_seconds FROM dual_floats;
CREATE TABLE
Time: 26.080 ms
INSERT 0 1000000
Time: 2689.349 ms (00:02.689)
```

```test
 Seq Scan on dual_floats  (cost=0.00..22810.00 rows=1000000 width=1) (actual time=0.022..100.542 rows=1000000 loops=1)
 Planning Time: 0.255 ms
 Execution Time: 132.750 ms
```

## Overlaps | framestamp_range w/indexes

#### Before

```sql
INSERT INTO framestamps_02 (id, a, b)
SELECT 
    gen_random_uuid(),
    (
        (generate_series(1, 100000), 24), 
        ((24, 1), '{}')
    )::framestamp,
    (
        (generate_series(100001, 200000), 24), 
        ((24, 1), '{}')
    )::framestamp;

EXPLAIN ANALYZE SELECT events_01.id, events_02.id FROM framestamps_02 AS events_01
INNER JOIN framestamps_02 AS events_02 
	ON framestamp_range(events_01.a, events_01.b) && framestamp_range(events_02.a, events_02.b)
	LIMIT 100000;
```

```text
 Limit  (cost=0.29..5531.23 rows=100000 width=32) (actual time=1.044..322.084 rows=100000 loops=1)
   ->  Nested Loop  (cost=0.29..24889264.00 rows=450000000 width=32) (actual time=1.044..317.477 rows=100000 loops=1)
         ->  Seq Scan on framestamps_02 events_01  (cost=0.00..7000.00 rows=300000 width=80) (actual time=0.014..0.014 rows=1 loops=1)
         ->  Index Scan using framestamps_a_b_range on framestamps_02 events_02  (cost=0.29..52.94 rows=3000 width=80) (actual time=0.983..310.935 rows=100000 loops=1)
               Index Cond: (framestamp_range(a, b) && framestamp_range(events_01.a, events_01.b))
 Planning Time: 0.745 ms
 Execution Time: 324.626 ms
```

#### After

```sql
INSERT INTO framestamps_02 (id, a, b)
SELECT 
    gen_random_uuid(),
    (
        generate_series(1, 100000),
        24,
        24,
        1,
        '{}'
    )::framestamp,
    (
        generate_series(100001, 200000),
        24,
        24,
        1,
        '{}'
    )::framestamp;

EXPLAIN ANALYZE SELECT events_01.id, events_02.id FROM framestamps_02 AS events_01
INNER JOIN framestamps_02 AS events_02 
	ON framestamp_range(events_01.a, events_01.b) && framestamp_range(events_02.a, events_02.b)
	LIMIT 100000;
```

```text
 Limit  (cost=0.29..5554.35 rows=100000 width=32) (actual time=1.951..194.535 rows=100000 loops=1)
   ->  Nested Loop  (cost=0.29..7719385.69 rows=138986128 width=32) (actual time=1.950..189.903 rows=100000 loops=1)
         ->  Seq Scan on framestamps_02 events_01  (cost=0.00..3890.25 rows=166725 width=80) (actual time=0.017..0.017 rows=1 loops=1)
         ->  Index Scan using framestamps_a_b_range on framestamps_02 events_02  (cost=0.29..29.61 rows=1667 width=80) (actual time=1.899..183.530 rows=100000 loops=1)
               Index Cond: (framestamp_range(a, b) && framestamp_range(events_01.a, events_01.b))
 Planning Time: 0.993 ms
 Execution Time: 197.044 ms
```