@peake100 2023-07-19
Thanks for taking a look! I look forward to the spellcheck PR and appreciate you doing the legwork there.